pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aecos"
version = "0.7.0"
description = "AEC OS — AI-assisted operating system for architecture, engineering, and construction"
requires-python = ">=3.11"
dependencies = [
    "ifcopenshell>=0.8",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
visualization = [
    "pygltflib>=1.0",
]
speckle = [
    "specklepy>=2.0",
]
sync = [
    "requests>=2.28",
]
notifications = [
    "requests>=2.28",
]
slack = [
    "slack-bolt>=1.18",
]
teams = [
    "botbuilder-core>=4.15",
]
training = [
    "axolotl>=0.4",
    "peft>=0.8",
    "accelerate>=0.25",
    "datasets>=2.0",
]
all = [
    "pygltflib>=1.0",
    "specklepy>=2.0",
    "requests>=2.28",
    "slack-bolt>=1.18",
    "botbuilder-core>=4.15",
    "axolotl>=0.4",
    "peft>=0.8",
    "accelerate>=0.25",
    "datasets>=2.0",
]

[tool.setuptools.packages.find]
include = ["aecos*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
